Moved teacher related code from ProjectService to TeacherProjectService. #2676
