+++ src/main/webapp/site/src/app/authoring-tool/add-component/choose-new-component-location/choose-new-component-location.component.spec.ts
@@ -2,8 +2,8 @@ import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { TestBed } from '@angular/core/testing';
 import { UpgradeModule } from '@angular/upgrade/static';
 import { ConfigService } from '../../../../../../wise5/services/configService';
-import { ProjectService } from '../../../../../../wise5/services/projectService';
 import { TeacherDataService } from '../../../../../../wise5/services/teacherDataService';
+import { TeacherProjectService } from '../../../../../../wise5/services/teacherProjectService';
 import { UtilService } from '../../../../../../wise5/services/utilService';
 import { ChooseNewComponentLocation } from './choose-new-component-location.component';
 
@@ -42,7 +42,7 @@ class MockUpgradeModule {
 }
 
 let component: ChooseNewComponentLocation;
-let projectService: ProjectService;
+let teacherProjectService: TeacherProjectService;
 
 describe('ChooseNewComponentLocation', () => {
   beforeEach(() => {
@@ -51,14 +51,14 @@ describe('ChooseNewComponentLocation', () => {
       providers: [
         ChooseNewComponentLocation,
         ConfigService,
-        { provide: ProjectService, useClass: MockProjectService },
+        { provide: TeacherProjectService, useClass: MockProjectService },
         { provide: TeacherDataService, useClass: MockTeacherDataService },
         { provide: UpgradeModule, useClass: MockUpgradeModule },
         UtilService
       ]
     });
     component = TestBed.inject(ChooseNewComponentLocation);
-    projectService = TestBed.inject(ProjectService);
+    teacherProjectService = TestBed.inject(TeacherProjectService);
   });
 
   it('should have nodeId and components set after initialization', () => {
@@ -68,10 +68,10 @@ describe('ChooseNewComponentLocation', () => {
   });
 
   it('insertComponentAfter() should create a new component and save the project', () => {
-    spyOn(projectService, 'createComponent').and.returnValue({ id: 'comp3', type: 'Discussion' });
-    spyOn(projectService, 'saveProject').and.returnValue(new Promise(() => {}));
+    spyOn(teacherProjectService, 'createComponent').and.returnValue({ id: 'comp3', type: 'Discussion' });
+    spyOn(teacherProjectService, 'saveProject').and.returnValue(new Promise(() => {}));
     component.insertComponentAfter('comp2');
-    expect(projectService.createComponent).toHaveBeenCalled();
-    expect(projectService.saveProject).toHaveBeenCalled();
+    expect(teacherProjectService.createComponent).toHaveBeenCalled();
+    expect(teacherProjectService.saveProject).toHaveBeenCalled();
   });
 });
