+++ src/main/webapp/site/src/app/preview/modules/choose-branch-path-dialog/choose-branch-path-dialog.component.ts
@@ -1,7 +1,5 @@
 import { Component, OnInit, Inject } from '@angular/core';
 import { MAT_DIALOG_DATA } from '@angular/material/dialog';
-import { NodeService } from '../../../../../../wise5/services/nodeService';
-import { ProjectService } from '../../../../../../wise5/services/projectService';
 
 @Component({
   selector: 'app-choose-branch-path-dialog',
